requirements.txt
torch==1.10.1
torch_geometric==2.0.4
rdkit-pypi==2023.3.3
biopython==1.79
pandas==1.3.3
scipy==1.7.1
requests==2.26.0
tqdm==4.64.0
numpy==1.21.2